+++ server/tutor-domain/src/main/java/cn/forever24/tutor/content/ContentImportStatus.java
@@ -1,0 +1,8 @@
+package cn.forever24.tutor.content;
+
+public enum ContentImportStatus {
+    RECEIVED,
+    REJECTED,
+    IMPORTED_DRAFT,
+    IMPORT_FAILED
+}